+++ src/_pytask/settings.py
@@ -1,15 +1,80 @@
 from __future__ import annotations
 
+import importlib.util
 from enum import Enum
+from pathlib import Path
+from typing import Iterable
 
+import click
 import typed_settings as ts
 from click import BadParameter
 from click import Context
+from pluggy import PluginManager
 from sqlalchemy.engine import URL
 from sqlalchemy.engine import make_url
 from sqlalchemy.exc import ArgumentError
 
+from _pytask.capture_utils import CaptureMethod
+from _pytask.capture_utils import ShowCapture
 from _pytask.click import EnumChoice
+from _pytask.path import import_path
+from _pytask.pluginmanager import hookimpl
+from _pytask.pluginmanager import register_hook_impls_from_modules
+from _pytask.pluginmanager import storage
+
+
+def _hook_module_callback(
+    ctx: Context,
+    name: str,  # noqa: ARG001
+    value: tuple[str, ...],
+) -> Iterable[str | Path]:
+    """Register the user's hook modules from the configuration file."""
+    if not value:
+        return value
+
+    parsed_modules = []
+    for module_name in value:
+        if module_name.endswith(".py"):
+            path = Path(module_name)
+            if ctx.params["config"]:
+                path = ctx.params["config"].parent.joinpath(path).resolve()
+            else:
+                path = Path.cwd().joinpath(path).resolve()
+
+            if not path.exists():
+                msg = (
+                    f"The hook module {path} does not exist. "
+                    "Please provide a valid path."
+                )
+                raise click.BadParameter(msg)
+            module = import_path(path, ctx.params["root"])
+            parsed_modules.append(module.__name__)
+        else:
+            spec = importlib.util.find_spec(module_name)
+            if spec is None:
+                msg = (
+                    f"The hook module {module_name!r} is not importable. "
+                    "Please provide a valid module name."
+                )
+                raise click.BadParameter(msg)
+            parsed_modules.append(module_name)
+
+    # If there are hook modules, we register a hook implementation to add them.
+    # ``pytask_add_hooks`` is a historic hook specification, so even command line
+    # options can be added.
+    if parsed_modules:
+
+        class HookModule:
+            @staticmethod
+            @hookimpl
+            def pytask_add_hooks(pm: PluginManager) -> None:
+                """Add hooks."""
+                register_hook_impls_from_modules(pm, parsed_modules)
+
+        pm = storage.get()
+        pm.register(HookModule)
+
+    return parsed_modules
 
 
 @ts.settings
@@ -145,6 +210,7 @@ class Markers:
         click={"param_decls": ["--strict-markers"], "is_flag": True},
         help="Raise errors for unknown markers.",
     )
+    markers: dict[str, str] = ts.option(factory=dict, click={"hidden": True})
     marker_expression: str = ts.option(
         default="",
         click={
@@ -166,4 +232,61 @@ class Settings:
         default=False,
         click={"param_decls": ("--debug-pytask",), "is_flag": True},
         help="Trace all function calls in the plugin framework.",
+    )
+    editor_url_scheme: str = ts.option(
+        default="file",
+        click={"param_decls": ["--editor-url-scheme"]},
+        help=(
+            "Use file, vscode, pycharm or a custom url scheme to add URLs to task "
+            "ids to quickly jump to the task definition. Use no_link to disable URLs."
+        ),
+    )
+    ignore: tuple[str, ...] = ts.option(
+        factory=tuple,
+        help=(
+            "A pattern to ignore files or directories. Refer to 'pathlib.Path.match' "
+            "for more info."
+        ),
+        click={"param_decls": ["--ignore"], "multiple": True},
+    )
+    verbose: int = ts.option(
+        default=1,
+        help="Make pytask verbose (>= 0) or quiet (= 0).",
+        click={
+            "param_decls": ["-v", "--verbose"],
+            "type": click.IntRange(0, 2),
+            "count": True,
+        },
+    )
+    hook_module: tuple[str, ...] = ts.option(
+        factory=list,
+        help="Path to a Python module that contains hook implementations.",
+        click={
+            "param_decls": ["--hook-module"],
+            "multiple": True,
+            "is_eager": True,
+            "callback": _hook_module_callback,
+        },
+    )
+    config_file: Path | None = None
+
+
+@ts.settings
+class Capture:
+    """Settings for capturing."""
+
+    capture: CaptureMethod = ts.option(
+        default=CaptureMethod.FD,
+        click={"param_decls": ["--capture"]},
+        help="Per task capturing method.",
+    )
+    s: bool = ts.option(
+        default=False,
+        click={"param_decls": ["-s"], "is_flag": True},
+        help="Shortcut for --capture=no.",
+    )
+    show_capture: ShowCapture = ts.option(
+        default=ShowCapture.ALL,
+        click={"param_decls": ["--show-capture"]},
+        help="Choose which captured output should be shown for failed tasks.",
     )
