+++ src/_pytask/task.py
@@ -23,7 +23,7 @@ if TYPE_CHECKING:
 @hookimpl
 def pytask_parse_config(config: Settings) -> None:
     """Parse the configuration."""
-    config["markers"]["task"] = (
+    config.markers.markers["task"] = (
         "Mark a function as a task regardless of its name. Or mark tasks which are "
         "repeated in a loop. See this tutorial for more information: "
         "[link https://bit.ly/3DWrXS3]https://bit.ly/3DWrXS3[/]."
