+++ tests/test_collect_utils.py
@@ -2,7 +2,7 @@ from __future__ import annotations
 
 import pytest
 from _pytask.collect_utils import _find_args_with_product_annotation
-from pytask import Product  # noqa: TCH002
+from pytask import Product
 from typing_extensions import Annotated
